+++ src/Shipment/Models/ShipmentStatusProxy.php
@@ -1,0 +1,21 @@
+<?php
+
+declare(strict_types=1);
+
+/**
+ * Contains the ShipmentStatusProxy class.
+ *
+ * @copyright   Copyright (c) 2022 Attila Fulop
+ * @author      Attila Fulop
+ * @license     MIT
+ * @since       2022-02-28
+ *
+ */
+
+namespace Vanilo\Shipment\Models;
+
+use Konekt\Concord\Proxies\EnumProxy;
+
+class ShipmentStatusProxy extends EnumProxy
+{
+}