+++ src/Shipment/Changelog.md
@@ -4,4 +4,4 @@
 ##### 2022-02-XX
 
 - The very first release of Shipment module (replaces the never-released "Shipping" module)
-- 
+- Shipment and Shipment status classes